PRSDK-7940: ipc-lld: Make Linux/C66x Vring Memory as Non-Cached
The Vring memory that is used for C66x cores communication with
A72 Linux needs to be set as Non-cached as this is shared memory
between the two cores. Without this, communication with Linux
on A72 did not happen properly when L1D/L2 cache was enabled
on the C66x cores. This was due to stale values in the
shared vring memory.

Additionally, to make sure the trace buffer is readable from
Linux debugfs trace, the trace buffer needs to be flushed
periodically when cache is enabled. Add a function that will
run when entering idle that will flush the contents of the
trace buffer to memory.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>