ipc-lld: qnx: fix issue with handling out of order buffers
The remote core may have 2 threads sending messages to the host.
In this case it is possible that thread A may get an available
buffer, and then thread B can come along and interrupt thread A
and also get an available buffer, and then put that buffer to
the used list before thead A can put the buffer to the used list.

In this case, the available buffer index will no longer match
the index of the buffer in the desc array. When this happens,
the same message gets read twice, and also the desc array itself
becomes corrupted due to the receiving-side always assuming that
the index will be matching. Instead, when getting and putting
buffers, a token should be taken in order to indicate the location
of the message in the desc array. When adding buffers back,
the desc array should only be updated at that token index. The id of
the buffer in the available array should also be updated to the token
value being added.

In this way, re-ordering of buffers is accounted for.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>