sciclient: Add element to request struct for forwarding status
Update the sciclient request parameters to add a field for forward
status to indicate whether that the message should be forwarded.

This is only intended to be set indirectly by sciserver, and also
sciclient_direct depending on the message type. For sciclient_indirect,
this will be forced to no-forward status to prevent any application from
erroneously setting. The status is always cleared after message
processing to prevent accidental reuse of the status upon next message
send.

This includes a hack to update sciclient request parameters which are
given the const qualifier. This is due to an API limitation, and is
considered acceptable based on convention elsewhere in sciclient to work
around the const through indirect dereferencing.

Signed-off-by: Stephen Molfetta <sjmolfetta@ti.com>