sciserver: Add config option to select boardcfg source
Board configuration may be built locally into the server-integrated
application within the PDK build environment. However, in the case that
the server is integrated standalone into a separate boot environment
(e.g. Linux SPL), board configurations are expected to be queued prior
to booting the server firmware.

A config option is provided to enable integrating applications to select
the boardcfg local structure or to bypass the local processing to enable
externally provided configurations over the sec proxy queue.

Signed-off-by: Stephen Molfetta <sjmolfetta@ti.com>