sciclient: register contextID for response ISR
Response ISR attempts to look up the context ID by the interrupt number.
This fails when multiple CPUs use the same local interrupt number for
different signals routed to their interrupt controllers, causing lookup
of the wrong context.

Instead, register the context ID directly with the ISR so that there is
no need to look up the context.

Fixes PDK-8086

Signed-off-by: Stephen Molfetta <sjmolfetta@ti.com>