sciclient: Do not force message forwarding for DMSC-only functions
RM PSIL, baseport, and secure are entirely processed on DMSC. mcu1_0
requests for these messages should not be treated as forwarded and
rather as native calls. This has an effect on which queue is used to
send the message, and the mcu1_0 queue has extended depth to enable SBL
sequencing for self-reset.

Signed-off-by: Stephen Molfetta <sjmolfetta@ti.com>