sciclient: pass request flags for PM set proxy
When PM set device is called, a proxy is used specifically for MCU
reboot due to inability to self-reset in proper sequence. However, it is
valid to set the power state without queueing commands in sequence,
which in that case, request flags must be honored. Otherwise, race
conditions may lead to unpredictable behavior.

Signed-off-by: Stephen Molfetta <sjmolfetta@ti.com>