sciserver: tirtos: convert semaphore to counting
Semaphore for scheduling the tasks needs to be counting. This is because
it is possible for both MAIN and MCU domain messages to arrive within a
small enough window of time where both ISRs trigger before the first
message is fully processed. A binary semaphore has the effect of
dropping the first interrupt in this series.

Signed-off-by: Stephen Molfetta <sjmolfetta@ti.com>