sciserver: Add check for softerror in isr
Soft error indicates that host ID does in the message header does not
match the host associated with the queue that the message was received.
This is not a fatal error, just results in the message being dropped. In
that case, re-enable the interrupt and proceed with the next message
when received.

Signed-off-by: Stephen Molfetta <sjmolfetta@ti.com>