sciclient: sciclient calls to initialize req param and resp param
As part of a strange debug, it was found the SBL was not working with TI
CLANG compiler. The nature of the failure was the DDR was not
initialized. As part of that debug it was found the EMIF PSC was not
getting enabled. Further digging led to the understanding that the PM
module was not getting the right host ID when trying to initialize the
PSC for EMIF in some cases. This was because the Sciclient API was not
initializing the value of the parameter for forwardMsg. In this
particular case the forwardMsg field in the uninitialized reqPrm caused
the Sciclient API to think this is a forwarded message and did not set
the host ID for the PM API. The PM API in turn left the EMIF PSC in a
disabled state leading to the DDR not being initialized.

The fix in this commit is to initialize the SciclientReq and Resp Params
to ensure when there is an additional parameter added to these
structures, they are not uninitialized.

Fixes: PDK-10821

Signed-off-by: Piyali Goswami <piyali_g@ti.com>