ETHFW-1926 - timesync: v2: cpsw: Fix offset adjustment direction
The adjustment value in this PTP stack implementation is calculated as
"local" - "remote" timestamps, hence a positive value means that local
timestamp has to be decremented.

However, a positive adjValInNsecs value passed to CPSW IOCTL means that
CPSW will add a corresponding PPM value to the timestap.  This can be
fixed by changing the sign of the adjustment value.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>