DM trace: Do trace configuration
Debug trace configuration information is present in the base board
config. But base board config is only given to TIFS not to DM by the
bootloader. But DM needs to know what is the debug trace configuration
present in the board config for configuring DM's trace. During boot time
get the debug trace configuration form TIFS and configure the DM's trace

Fixes: SYSFW-5366

Signed-off-by: sebin francis <sebin.francis@ti.com>