am62a: remove lpm components from linker
In am62x and am62a few section of code has different run and load
address. Because of this there is hack added in the self_reset that
assumes the load and run address will always differ a constant value.
But this is not ture for am62a. lpm components are not enabled in am62a,
because of this the lpm section are empty. this creates a hole in the
memory and break the assumption that load and run address of each
section is always same. Remove the lpm section from the linker to fix
this.

Signed-off-by: sebin francis <sebin.francis@ti.com>