[PSDKQAS-793] pdk: ipc: RPMessage_recv: do not leave dangling pointer on error return
In case of no incoming message already available available at the call
time, RPMessage_recv() saves the data pointer passed to it, for use in
incoming message handler, and then sleeps.

This sleep can be terminated by timeout or other error condition.
In this case, the saved pointer was left. Later, on arrive of incoming
message, that messagfe was copied to the memory pointed by that pointer,
causing overwrite of whatever is in that memory, and unpredictable
consequences.

Fix it by clearing the pointer. Be careful to handle possible race if
incoming message arrives simultaneously with the error path being
executed.

Signed-off-by: Praveen Rao <prao@ti.com>