Allow to switch between IORET and PMIC only mode
The firmware is not supposed to support the both case in the same time
on the EVM. So a choice has to be made at compilation time using the
PMIC_WAKEUP define.

However even when PMIC_WAKEUP is not defined the EVM still managed to
wake-up from the SW12 (CAN0_WAKE). In this configuration, when doing
continuous suspend, the EVM alternate between being automatically
wake up and being wake up from the SW12.

Signed-off-by: Gregory CLEMENT <gregory.clement@bootlin.com>