[QNX][UDMA] Added an attribute to track per process proxy allocations
The udma driver object uses a proxy in case the associated ring is configured
in MESSAGE mode. For A72, due to lack of enough dedicated proxies, we allocate
dedicate proxies per udma client/process. In case the number of clients
are more than 3 (with the current resource partitioning for AHP), the allocation
will fail and this could cause the clients that come up last to fail.

Signed-off-by: Subbu Chanderashekarapuram <s-chanderashekarapuram@ti.com>