[QNX][UDMA] Add support to use the ring specific proxy for queuing
The ring specific proxy is assigned for QNX such that each process
that uses a ring can use a dedicate proxy to communicate with the
underlying ring when operating in message mode. Without this change,
the APIs would fallback to using the default global proxy thread (proxy 0)
for transfers. This was causing issues, since the access to this register
space is not arbirated system wide and could cause corruption of
the dma message. This usually manifests as a SIGBUS error.

Signed-off-by: Subbu Chanderashekarapuram <s-chanderashekarapuram@ti.com>