[QNX][OSAL] Updates for SDP 8.0 compatibility
Updated the delay funtion to use an API that is available
in both SDP 7.1 and SDP 8.0

The interrupt Mask and Unmask APIs accept 2 parameters. The first
one being the interruptNum and the second being eventId. Please
refer to the InterruptMask()/InterruptUnmask() in the QNX docs
for API documentation. Until SDP 7.1 this API was happy to accept
a -1 as input for the id parameter. However, with SDP 8.0 the
API behaves as intended. Hence updated the implementation to reflect
the same.

Also implemented a helper function to fetch the hwi datastructure
corresponding to the interrupt number.

NOTE: The logic to traverse the global hwi assumes that the
structure is not manipulated when the traversal happens. This might
not be always true. It is prudent to add protection whenever accessing
the data structure.

Signed-off-by: Subbu Chanderashekarapuram <s-chanderashekarapuram@ti.com>