Remove NULL checks for functions in sciclient_boardcfg.c file
Sciclient boardcfg functions should retain the default boardcfg values
if the passed argument is NULL otherwise the boardcfg values are set
with the user passed values. Hence the recent addition of NULL checks
in sciclient boardcfg functions is removed.

Corresponding changes are done in the testcase for boardcfg functions
in sciclient extended unit testapp.

Signed-off-by: Thejasvi Konduru <t-konduru@ti.com>