[SCICLIENT]:Restructure sciclient APIs
Majority of the highlevel sciclient APIs have a request parameter, response
parameter as arguments. Request parameter contains the information regarding
the messageType, reqPayload and reqPayloadSize. The request is sent to
sciclient_service() which then executes the logic corresponding to the given
message type. The response parameter gets populated in sciclient_service()
with the response flags, respPayload and respPayloadSize.

Sciclient API returns PASS/FAIL based on the values returned from the
sciclient_service() and the value of the response flag.

The above check is converted to a more simpler form in all the SCICLIENT APIs.