lwip: Removal of lwipools.h per OS, lwippools is structured per SOC now.
Previously, lwipools configuration parameters were structured per OS.
So all platforms had one lwipools.h header file for each supported OS.

The right structure is to split the configuration parameters in SoC-dependent.
For a given SoC, a single SoC header is maintained (independent of OS).

The file organization in the new structure is:
- lwip-port/config/<SOC>/lwippools.h

Fixes: ETHFW-2736

Signed-off-by: Vaibhav Jindal <v-jindal@ti.com>