[PDK-13507][Sciclient]: Fix firewall exception notification testcase in sciclient unit testapp
Firewall exception notification testcase tries to access a protected address in DMSC region
and the DMSC sends a firewall exception interrupt to the host. On receiving the interrupt,
the corresponding ISR gets executed on the host and the associated ISR variable gets
incremented. Two such firewall exceptions occur in this testcase. The success/failure of the
testcase depends on the value of the ISR variable.

Following changes are made in the testapp
1.Changed the ISR variable to a volatile variable inorder to avoid the effect of compiler
optimizations.
2.The ISR variable is getting incremented correctly but the check on the value of the ISR vairable
is not happening correctly. This would have been due to the compiler trying to reorder read and write
operations corresponding to ISR variable in the assembly code. Hence used __sync_synchronize()
inorder to make sure that all the write operations on the ISR variable occur before the read
operation on the ISR variable in the main code.
3.On observing the TIFS firewall exception logs, the execption logging registers corresponding
to this testcase belongs to the address range (0x44000000U - 0x45FFFFFFU) and for this address
range, the local address value in M3/M4 and global address value is same. Hence a firewall
exception register which is out of this range is removed from the array of firewall exception
logging registers.

The testcase is passing on all the SoC's with the above changes.

Signed-off-by: Thejasvi Konduru <t-konduru@ti.com>